# docs/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: LLM Start Backend API
  version: "1.0.0"
  description: |
    Черновик публичного HTTP API v1. После появления приложения FastAPI источником правды становится сгенерированная схема
    (`/openapi.json`); этот файл — контрактная опора до и на время задач 03–05.
servers:
  - url: https://api.example.com
    description: Production (пример)
  - url: http://127.0.0.1:8000
    description: Локальная разработка
...
paths:
...
  /api/v1/cohorts/{cohort_id}/dialogues/messages:
    post:
      tags: [dialogues]
      summary: Отправить сообщение в диалог (создать или продолжить)
      description: |
        Создаёт диалог при первом обращении или продолжает существующий. Сохраняет реплики пользователя и ассистента
        согласно доменной модели (`Dialogue`, `DialogueTurn` в БД как пара Q/A), **если поток с `cohort_id` есть в БД**.
        Если поток не найден, ошибки нет: ответ в том же формате, контекст ведётся в памяти процесса (как
        `POST /api/v1/assistant/guest/messages`), персистентности в БД нет. Вызов LLM выполняется только на сервере.
      operationId: postCohortDialogueMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDialogueMessageRequest"
      responses:
        "200":
          description: Ответ ассистента с идентификаторами сообщений
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDialogueMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "502":
          $ref: "#/components/responses/BadGatewayLlm"
        "503":
          $ref: "#/components/responses/ServiceUnavailableLlm"
  /api/v1/dialogues/{dialogue_id}/reset:
    post:
      tags: [dialogues]
      summary: Сбросить контекст диалога
      description: |
        Аналог команды `/reset` в боте. Идемпотентность на уровне продукта — повторный сброс не ломает состояние;
        сервер может архивировать текущий диалог и создавать новый при следующем сообщении.
      operationId: postDialogueReset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DialogueId"
      responses:
        "204":
          description: Контекст сброшен
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/cohorts/{cohort_id}/progress-checkpoints:
    get:
      tags: [progress]
      summary: Список этапов прогресса потока
      operationId: listProgressCheckpoints
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
      responses:
        "200":
          description: Упорядоченный список этапов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressCheckpointListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/cohorts/{cohort_id}/memberships/{membership_id}/progress-records/{checkpoint_id}:
    put:
      tags: [progress]
      summary: Создать или обновить отметку прогресса по этапу
      description: |
        Upsert по паре (участие, этап). Соответствует сущности `ProgressRecord` в data-model.
      operationId: putProgressRecord
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
        - $ref: "#/components/parameters/MembershipId"
        - $ref: "#/components/parameters/CheckpointId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutProgressRecordRequest"
      responses:
        "200":
          description: Отметка сохранена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecordResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/cohorts/{cohort_id}/summary:
    get:
      tags: [cohorts]
      summary: Агрегированный срез по потоку (преподаватель)
      description: |
        Обзор участников и прогресса по этапам в рамках одного потока. Минимальный набор полей в v1;
        расширение — без ломки контракта (новые поля).
      operationId: getCohortSummary
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
        - $ref: "#/components/parameters/ViewerMembershipId"
      responses:
        "200":
          description: Сводка по потоку
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/cohorts/{cohort_id}/teacher-dashboard:
    get:
      tags: [cohorts]
      summary: Панель преподавателя (агрегат по потоку)
      description: |
        KPI с дельтой к предыдущей календарной неделе, ряд вопросов к ассистенту по дням, ленты Q&A и сдач, матрица прогресса.
        Реализовано в backend.
      operationId: getCohortTeacherDashboard
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
        - $ref: "#/components/parameters/ViewerMembershipId"
        - $ref: "#/components/parameters/ActivityDays"
        - $ref: "#/components/parameters/TurnsCursor"
        - $ref: "#/components/parameters/TurnsLimit"
        - $ref: "#/components/parameters/TurnsSearchQ"
      responses:
        "200":
          description: Данные дашборда
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/cohorts/{cohort_id}/leaderboard:
    get:
      tags: [cohorts]
      summary: Лидерборд потока (таблица и scatter)
      description: |
        Упорядоченный рейтинг и поля для осей scatter (уроки vs домашки). Реализовано в backend.
      operationId: getCohortLeaderboard
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
        - $ref: "#/components/parameters/ViewerMembershipIdOptional"
      responses:
        "200":
          description: Лидерборд
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/cohorts/{cohort_id}/memberships/{membership_id}/progress-overview:
    get:
      tags: [progress]
      summary: Кабинет студента — чекпоинты и свои записи прогресса
      description: |
        Участие должно принадлежать `cohort_id`; роль — student (MVP с общим Bearer).
      operationId: getMembershipProgressOverview
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CohortId"
        - $ref: "#/components/parameters/MembershipId"
      responses:
        "200":
          description: Обзор прогресса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgressOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/dialogues/{dialogue_id}/turns:
    get:
      tags: [dialogues]
      summary: История ходов диалога (Q/A)
      description: |
        Список `DialogueTurn` для канала веб/бот; пагинация по времени (`before_asked_at`, `limit`).
      operationId: listDialogueTurns
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DialogueId"
        - $ref: "#/components/parameters/TurnsLimit"
        - $ref: "#/components/parameters/BeforeAskedAt"
      responses:
        "200":
          description: Страница ходов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DialogueTurnsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        MVP: токен клиента (бот/веб) или внутренний ключ — см. `docs/tasks/impl/.../task-02-contracts/plan.md`.
  parameters:
    CohortId:
      name: cohort_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DialogueId:
      name: dialogue_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MembershipId:
      name: membership_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CheckpointId:
      name: checkpoint_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ViewerMembershipId:
      name: viewer_membership_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: |
        Идентификатор участия (`CohortMembership`) вызывающей стороны в этом потоке.
        Для `GET .../summary` должен указывать на участие с ролью преподавателя.
        При общем Bearer-токене клиента (MVP) без JWT это обязательный способ указать, кто запрашивает сводку.
    ViewerMembershipIdOptional:
      name: viewer_membership_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Опционально: если указан — проверяется, что участие принадлежит этому потоку. Иначе данные доступны без проверки зрителя (MVP).
    ActivityDays:
      name: activity_days
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 366
        default: 14
      description: Число последних календарных дней для ряда `activity_by_day` на дашборде.
    TurnsCursor:
      name: turns_cursor
      in: query
      required: false
      schema:
        type: string
        nullable: true
      description: Непрозрачный курсор пагинации ленты ходов на дашборде (следующая страница).
    TurnsLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Размер страницы для ленты ходов или `GET .../dialogues/.../turns`.
    TurnsSearchQ:
      name: q
      in: query
      required: false
      schema:
        type: string
        nullable: true
      description: Подстрока поиска по тексту вопроса студента (ленты на дашборде).
    BeforeAskedAt:
      name: before_asked_at
      in: query
      required: false
      schema:
        type: string
        format: date-time
        nullable: true
      description: Для пагинации истории — только ходы с `asked_at` строго раньше значения.
  schemas:
    ClientChannel:
      type: string
      enum: [telegram, web]
      description: Канал клиента (диалог в data-model)
    PostDialogueMessageRequest:
      type: object
      required:
        - membership_id
        - channel
        - content
      properties:
        membership_id:
          type: string
          format: uuid
          description: Идентификатор `CohortMembership` инициатора
        dialogue_id:
          type: string
          format: uuid
          nullable: true
          description: Если null — сервер выбирает/создаёт активный диалог для пары (участие, канал)
        channel:
          $ref: "#/components/schemas/ClientChannel"
        content:
          type: string
          description: Текст пользователя (не логировать в открытом виде на стороне клиента в проде)
    PostDialogueMessageResponse:
      type: object
      required:
        - dialogue_id
        - user_message_id
        - assistant_message
      properties:
        dialogue_id:
          type: string
          format: uuid
        user_message_id:
          type: string
          format: uuid
        assistant_message:
          type: object
          required:
            - id
            - content
            - created_at
          properties:
            id:
              type: string
              format: uuid
            content:
              type: string
            created_at:
              type: string
              format: date-time
    ProgressCheckpointItem:
      type: object
      required:
        - id
        - code
        - title
        - sort_order
        - required
        - is_homework
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        sort_order:
          type: integer
        required:
          type: boolean
        is_homework:
          type: boolean
          description: Участие этапа в метриках «домашки» (лидерборд, KPI дашборда).
    ProgressCheckpointListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProgressCheckpointItem"
    ProgressStatus:
      type: string
      enum: [not_started, in_progress, completed, skipped]
      description: Статус фиксации прогресса (согласован с `data-model.md`)
    PutProgressRecordRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ProgressStatus"
        comment:
          type: string
          nullable: true
          maxLength: 2000
        submission_links:
          type: array
          items:
            type: string
          maxItems: 32
          description: URL сдачи (отдельно от текстового комментария).
    ProgressRecordResponse:
      type: object
      required:
        - id
        - cohort_id
        - membership_id
        - checkpoint_id
        - status
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        cohort_id:
          type: string
          format: uuid
        membership_id:
          type: string
          format: uuid
        checkpoint_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProgressStatus"
        comment:
          type: string
          nullable: true
        submission_links:
          type: array
          items:
            type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
...
    KpiWeekPair:
      type: object
      required:
        - current_week
        - previous_week
      properties:
        current_week:
          type: integer
          minimum: 0
        previous_week:
          type: integer
          minimum: 0
    TeacherDashboardKpis:
      type: object
      required:
        - active_students
        - homework_completed_events
        - avg_progress_percent
        - dialogue_questions
      properties:
        active_students:
          $ref: "#/components/schemas/KpiWeekPair"
        homework_completed_events:
          $ref: "#/components/schemas/KpiWeekPair"
          description: Число завершённых домашних чекпоинтов за неделю (сумма по группе).
        avg_progress_percent:
          type: object
          required:
            - current_week
            - previous_week
          properties:
            current_week:
              type: number
              format: float
              minimum: 0
              maximum: 100
            previous_week:
              type: number
              format: float
              minimum: 0
              maximum: 100
        dialogue_questions:
          $ref: "#/components/schemas/KpiWeekPair"
          description: Число новых вопросов (ходов) по потоку за календарную неделю, все каналы.
    ActivityByDayItem:
      type: object
      required:
        - day
        - question_count
      properties:
        day:
          type: string
          format: date
        question_count:
          type: integer
          minimum: 0
    DashboardDialogueTurnItem:
      type: object
      required:
        - membership_id
        - user_message_id
        - assistant_message_id
        - student_display_name
        - question_text
        - answer_text
        - asked_at
      properties:
        membership_id:
          type: string
          format: uuid
        user_message_id:
          type: string
          format: uuid
        assistant_message_id:
          type: string
          format: uuid
        student_display_name:
          type: string
          nullable: true
        question_text:
          type: string
        answer_text:
          type: string
        asked_at:
          type: string
          format: date-time
    DashboardRecentSubmissionItem:
      type: object
      required:
        - membership_id
        - checkpoint_id
        - status
        - updated_at
      properties:
        membership_id:
          type: string
          format: uuid
        student_display_name:
          type: string
          nullable: true
        checkpoint_id:
          type: string
          format: uuid
        checkpoint_title:
          type: string
        status:
          $ref: "#/components/schemas/ProgressStatus"
        comment:
          type: string
          nullable: true
        submission_links:
          type: array
          items:
            type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
    DashboardMatrixCell:
      type: object
      required:
        - checkpoint_id
        - status
      properties:
        checkpoint_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProgressStatus"
        updated_at:
          type: string
          format: date-time
          nullable: true
    DashboardMatrixRow:
      type: object
      required:
        - membership_id
        - display_name
        - score_completed
        - score_total
        - cells
      properties:
        membership_id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        score_completed:
          type: integer
          minimum: 0
        score_total:
          type: integer
          minimum: 0
        cells:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMatrixCell"
    TeacherDashboardRecentTurns:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardDialogueTurnItem"
        next_cursor:
          type: string
          nullable: true
    TeacherDashboardResponse:
      type: object
      required:
        - cohort_id
        - kpis
        - activity_by_day
        - recent_turns
        - recent_submissions
        - matrix
      properties:
        cohort_id:
          type: string
          format: uuid
        cohort_title:
          type: string
          nullable: true
        kpis:
          $ref: "#/components/schemas/TeacherDashboardKpis"
        activity_by_day:
          type: array
          items:
            $ref: "#/components/schemas/ActivityByDayItem"
        recent_turns:
          $ref: "#/components/schemas/TeacherDashboardRecentTurns"
        recent_submissions:
          type: array
          items:
            $ref: "#/components/schemas/DashboardRecentSubmissionItem"
        matrix:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMatrixRow"
    LeaderboardCheckpointStatus:
      type: object
      required:
        - checkpoint_id
        - status
      properties:
        checkpoint_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProgressStatus"
    LeaderboardEntry:
      type: object
      required:
        - rank
        - membership_id
        - user_id
        - progress_percent
        - completed_checkpoints
        - total_checkpoints
        - homework_completed
        - lesson_completed
        - per_checkpoint
      properties:
        rank:
          type: integer
          minimum: 1
        membership_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        progress_percent:
          type: number
          format: float
          minimum: 0
          maximum: 100
        completed_checkpoints:
          type: integer
          minimum: 0
        total_checkpoints:
          type: integer
          minimum: 0
        homework_completed:
          type: integer
          minimum: 0
        lesson_completed:
          type: integer
          minimum: 0
          description: Завершённые этапы без признака ДЗ (для горизонтали scatter).
        scatter_x:
          type: number
          format: float
          nullable: true
          description: Нормализованная координата по урокам (0–1 или доля — как в реализации).
        scatter_y:
          type: number
          format: float
          nullable: true
          description: Нормализованная координата по домашкам.
        per_checkpoint:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardCheckpointStatus"
    LeaderboardResponse:
      type: object
      required:
        - cohort_id
        - checkpoints
        - entries
      properties:
        cohort_id:
          type: string
          format: uuid
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/ProgressCheckpointItem"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
    StudentProgressRecordItem:
      type: object
      required:
        - checkpoint_id
        - status
      properties:
        checkpoint_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProgressStatus"
        updated_at:
          type: string
          format: date-time
          nullable: true
        comment:
          type: string
          nullable: true
        submission_links:
          type: array
          items:
            type: string
          nullable: true
    StudentProgressOverviewResponse:
      type: object
      required:
        - cohort_id
        - membership_id
        - checkpoints
        - records
      properties:
        cohort_id:
          type: string
          format: uuid
        membership_id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/ProgressCheckpointItem"
        records:
          type: array
          items:
            $ref: "#/components/schemas/StudentProgressRecordItem"
    DialogueTurnListItem:
      type: object
      required:
        - user_message_id
        - assistant_message_id
        - question_text
        - answer_text
        - asked_at
        - answered_at
      properties:
        user_message_id:
          type: string
          format: uuid
        assistant_message_id:
          type: string
          format: uuid
        question_text:
          type: string
        answer_text:
          type: string
        asked_at:
          type: string
          format: date-time
        answered_at:
          type: string
          format: date-time
    DialogueTurnsListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DialogueTurnListItem"
        next_before_asked_at:
          type: string
          format: date-time
          nullable: true
          description: Передать как `before_asked_at` для следующей страницы (если не null).
    CohortSummaryParticipant:
      type: object
      required:
        - membership_id
        - user_id
        - role
      properties:
        membership_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [student, teacher]
        name:
          type: string
          nullable: true
        progress:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProgressStatus"
          description: Ключ — `checkpoint_id` (UUID строкой), значение — статус
    CohortSummaryResponse:
      type: object
      required:
        - cohort_id
        - checkpoints
        - participants
      properties:
        cohort_id:
          type: string
          format: uuid
        cohort_title:
          type: string
          nullable: true
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/ProgressCheckpointItem"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/CohortSummaryParticipant"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              description: Опционально — поля валидации или машиночитаемые детали
              oneOf:
                - type: array
                  items:
                    type: object
                - type: object
  responses:
    Unauthorized:
      description: Нет или неверный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Доступ запрещён для данного участника/роли
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Конфликт домена (например, нарушение уникальности при альтернативной семантике POST)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Тело не проходит схему (в FastAPI часто 422)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGatewayLlm:
      description: Ответ провайдера LLM некорректен или шлюз недоступен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: LLM_BAD_GATEWAY
              message: Assistant is temporarily unavailable
    ServiceUnavailableLlm:
      description: Провайдер LLM недоступен (таймаут, перегруз)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: LLM_UNAVAILABLE
              message: Assistant is temporarily unavailable